requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.6
cryptography==38.0.0
Flask==2.3.2
Flask-RESTful==0.3.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
six==1.16.0
Unidecode==1.3.6
urllib3==1.26.16
Werkzeug==2.3.6
